moved getDefaultHospitalPreposition in controller
